vectorscope: make into separate utility module
There are two sorts of scopes:

- Histogram and waveform primarily show per-channel lightness
  information.
- Vectorscope shows chromaticity information without regard to
  lightness.

A long-time request has been to show both at once. That way it is
possible quickly to get a sense of both the lightness and chromaticity
of an image without flipping between scopes.

This commit is one way to handle this: It makes a new "vectorscope"
module which will only show vectorscope. The new module can be
displayed along with the existing histogram/waveform module.
